backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
librosa==0.10.1
soundfile==0.12.1
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.2
huggingface_hub==0.19.4
mediapipe==0.10.7
scipy==1.11.4
scikit-learn==1.3.2